move App back to src/
